relax constraint on nokogiri version dependency
It's understood that the maintainers of this gem want to support older
versions of ruby, which means that users of ruby 1.9 and 2.0 have to use
a pre-1.7.0 version of nokogiri.

However, users of modern rubies should be able to take advantage of the
features and security patches available in nokogiri 1.7.0 and later.

This patch relaxes the constraint so that any version of nokogiri on
1.6.x and 1.7.x release branches is acceptable. Users of the library can
then rely on bundler (via nokogiri's gemspec) to determine the latest
supported version of nokogiri that works for their ruby interpreter.

Note that older version of bundler may not use gems'
`required_ruby_version` declaration to pick the right version. Based on
a cursory examination of Bundler's changelog:

  https://github.com/bundler/bundler/blob/master/CHANGELOG.md

it appears as though bundler 1.13.0.rc.2 (2016-08-21) might be the first
version to respect this constraint.